+++ examples/template-hydrogen-default/src/pages/pages/[handle].server.jsx
@@ -1,9 +1,8 @@
-import {useShopQuery, RawHtml} from '@shopify/hydrogen';
+import {useShopQuery, RawHtml, Seo} from '@shopify/hydrogen';
 import gql from 'graphql-tag';
 
 import Layout from '../../components/Layout.server';
 import NotFound from '../../components/NotFound.server';
-import PageSeo from '../../components/PageSeo.client';
 
 export default function Page({params}) {
   const {handle} = params;
@@ -17,7 +16,7 @@ export default function Page({params}) {
 
   return (
     <Layout>
-      <PageSeo page={page} />
+      <Seo page={page} />
       <h1 className="text-2xl font-bold">{page.title}</h1>
       <RawHtml string={page.body} className="prose mt-8" />
     </Layout>