+++ examples/template-hydrogen-default/src/pages/products/[handle].server.jsx
@@ -1,7 +1,6 @@
-import {useShopQuery, ProductProviderFragment} from '@shopify/hydrogen';
+import {useShopQuery, ProductProviderFragment, Seo} from '@shopify/hydrogen';
 import gql from 'graphql-tag';
 
-import ProductSeo from '../../components/ProductSeo.client';
 import ProductDetails from '../../components/ProductDetails.client';
 import NotFound from '../../components/NotFound.server';
 import Layout from '../../components/Layout.server';
@@ -10,12 +9,7 @@ export default function Product({country = {isoCode: 'US'}, params}) {
   const {handle} = params;
 
   const {
-    data: {
-      product,
-      shop: {
-        primaryDomain: {url: shopUrl},
-      },
-    },
+    data: {product},
   } = useShopQuery({
     query: QUERY,
     variables: {
@@ -30,7 +24,7 @@ export default function Product({country = {isoCode: 'US'}, params}) {
 
   return (
     <Layout>
-      <ProductSeo product={product} shopUrl={shopUrl} />
+      <Seo product={product} />
       <ProductDetails product={product} />
     </Layout>
   );
@@ -49,11 +43,6 @@ const QUERY = gql`
     $numProductSellingPlanGroups: Int = 0
     $numProductSellingPlans: Int = 0
   ) @inContext(country: $country) {
-    shop {
-      primaryDomain {
-        url
-      }
-    }
     product: product(handle: $handle) {
       id
       description
@@ -66,6 +55,9 @@ const QUERY = gql`
         edges {
           node {
             url
+            height
+            width
+            altText
           }
         }
       }