+++ packages/hydrogen/src/components/Seo/ProductSeo.client.tsx
@@ -1,0 +1,104 @@
+import React from 'react';
+import {Helmet} from 'react-helmet-async';
+
+import {useServerRequest} from '../../foundation';
+
+import {TitleSeo} from './TitleSeo.client';
+import {DescriptionSeo} from './DescriptionSeo.client';
+import {TwitterSeo} from './TwitterSeo.client';
+import {ImageSeo} from './ImageSeo.client';
+
+import {Product} from './types';
+
+export function ProductSeo({
+  title,
+  description,
+  seo,
+  handle,
+  vendor,
+  images,
+  variants,
+}: Product) {
+  const seoTitle = seo?.title ?? title;
+  const seoDescription = seo?.description ?? description;
+
+  const {url} = useServerRequest();
+
+  let firstVariantPrice;
+  let firstImage;
+
+  const productSchema = {
+    '@context': 'http://schema.org/',
+    '@type': 'Product',
+    name: title,
+    description,
+    brand: {
+      '@type': 'Thing',
+      name: vendor,
+    },
+    url,
+  } as any;
+
+  if (images.edges.length > 0) {
+    firstImage = images.edges[0]?.node;
+    productSchema.image = firstImage.url;
+  }
+
+  if (variants.edges.length > 0) {
+    const firstVariant = variants.edges[0].node;
+    firstVariantPrice = firstVariant.priceV2;
+
+    if (firstVariant && firstVariant.sku) {
+      productSchema.sku = firstVariant.sku;
+    }
+
+    productSchema.offers = variants.edges.map(({node}) => {
+      const offerSchema = {
+        '@type': 'Offer',
+        availability: `https://schema.org/${
+          node.availableForSale ? 'InStock' : 'OutOfStock'
+        }`,
+        price: node.priceV2.amount,
+        priceCurrency: node.priceV2.currencyCode,
+      } as any;
+
+      if (node.sku) {
+        offerSchema.sku = node.sku;
+      }
+
+      if (node.image && node.image.url) {
+        offerSchema.image = node.image.url;
+      }
+
+      return offerSchema;
+    });
+  }
+
+  return (
+    <>
+      <Helmet>
+        <meta property="og:type" content="og:product" />
+        {firstVariantPrice && (
+          <meta
+            property="og:price:amount"
+            content={`${firstVariantPrice.amount}`}
+          />
+        )}
+        {firstVariantPrice && (
+          <meta
+            property="og:price:currency"
+            content={firstVariantPrice.currencyCode}
+          />
+        )}
+
+        <script type="application/ld+json">
+          {JSON.stringify(productSchema)}
+        </script>
+      </Helmet>
+      <TitleSeo title={seoTitle} />
+      <DescriptionSeo description={seoDescription} />
+      <TwitterSeo title={seoTitle} description={seoDescription} />
+      {firstImage && <ImageSeo {...firstImage} />}
+    </>
+  );
+}