+++ packages/hydrogen/src/components/Seo/examples/seo.example.tsx
@@ -1,0 +1,26 @@
+import {Seo, useShopQuery} from '@shopify/hydrogen';
+import gql from 'graphql-tag';
+
+const QUERY = gql`
+  query PageDetails($handle: String!) {
+    page(handle: $handle) {
+      title
+      body
+      seo {
+        title
+        description
+      }
+    }
+  }
+`;
+
+export function Page({params}) {
+  const {handle} = params;
+  const {data} = useShopQuery({query: QUERY, variables: {handle}});
+
+  if (data && data.page) {
+    return <Seo page={data.page} />;
+  }
+
+  return null;
+}